refactor(server): split http module and drop handler suffixes
